Add Abstract

Adds the abstract metadata element to a dataset list according to EML standards.

add_abstract(parent_element, abstract_file)

Arguments

parent_element

A list representing the EML project or dataset.

abstract_file

file path to abstract (word or markdown): see create_abstract

Value

The dataset or project with abstract appended

Examples

abstract_docx <- system.file("extdata", "Banet-Example", "metadata","abstract.docx", package = "EMLaide", mustWork = TRUE)  

dataset <- list() %>%
    add_abstract(abstract_docx)
dataset
#> $abstract
#> $abstract$section
#> $abstract$section[[1]]
#> [1] "<title>Overview</title>\n<para>\n      The Central Valley Project Improvement Act (CVPIA) funds habitat\n      improvement work in the Central Valley of California to increase\n      salmonid populations in furtherance of meeting CVPIA fish doubling\n      goals. This data package contains five datasets.\n    </para>"
#> 
#> $abstract$section[[2]]
#> [1] "<title>Enclosure Study – Growth Data</title>\n<para>\n      This dataset covers enclosure studies that examined salmonid\n      growth rates in the Sacramento River and focused on assessing\n      effectiveness of salmonid habitat improvement projects. Data was\n      collected in July and August 2019 from project sites, constructed\n      habitat project sites, and control sites where no treatment is\n      planned. Six enclosures with juvenile Fall Run Chinook salmon from\n      Coleman National Fish Hatchery were placed in each habitat type.\n      Fish growth was tracked for approximately 6.5 weeks. Annual\n      reports summarize the survey findings.\n    </para>"
#> 
#> $abstract$section[[3]]
#> [1] "<title>Enclosure Study – Gut Contents Data</title>\n<para>\n      This dataset covers enclosure studies that examined salmonid\n      growth rates in the Sacramento River and focused on assessing\n      effectiveness of salmonid habitat improvement projects. Data was\n      collected in July and August 2019 from project sites, constructed\n      habitat project sites, and control sites where no treatment is\n      planned. Six enclosures with juvenile Fall Run Chinook salmon from\n      Coleman National Fish Hatchery were placed in each habitat type.\n      Enclosures remained in the river for approximately 6.5 weeks.  At\n      the end of the study, fish were euthanized, and we dissected their\n      guts and enumerated the taxa found. Annual reports summarize the\n      survey findings.\n    </para>"
#> 
#> $abstract$section[[4]]
#> [1] "<title>Microhabitat Use Data</title>\n<para>\n      This dataset covers salmonid microhabitat use conducted in the\n      Sacramento River and focused on assessing effectiveness of\n      salmonid habitat improvement projects.  Surveys are conducted\n      roughly monthly and include pre-project sites, constructed habitat\n      project sites, and control sites where no treatment is planned. \n      Based upon habitat inventory data, annually identify which habitat\n      units within each side channel will be selected for the collection\n      of habitat-use data. Habitats are randomly selected but stratified\n      to include the full range of available habitat types to capture\n      the range in depths and velocities present, at approximately equal\n      surface areas each. Habitat selection attempts to capture the full\n      range of microhabitat cover types identified from microhabitat\n      mapping efforts. Snorkel surveys are conducted to achieve an\n      abundance index, with the addition that when a selected habitat\n      unit is encountered, the location of fish observed is marked with\n      a weighted flag on the stream bottom. The species / run, and size\n      of juvenile(s) and numbers observed at that location is recorded.\n      After the selected habitat unit has been completely surveyed,\n      flagged locations are revisited, and we collect habitat data\n      specific to each of those sample points.\n    </para>"
#> 
#> $abstract$section[[5]]
#> [1] "<title>Seining Data</title>\n<para>\n      This dataset covers salmonid lengths, weights, and conditions\n      collected via seining in the Sacramento River and focuses on\n      assessing effectiveness of salmonid habitat improvement projects. \n      Surveys are conducted roughly monthly and include pre-project\n      sites, constructed habitat project sites, and control sites where\n      no treatment is planned.  Wandering pole seine and/or simple arc\n      set beach seine methods were used to capture juvenile salmonids.\n      Randomized sample locations within side channels for wandering\n      pole seining were stratified by habitat type so that the range of\n      available habitat type and quality were sampled. Beach seining\n      methods were implemented at fixed sites. Mainstem sites were\n      selected far enough away from side channels so that juvenile fish\n      are not likely to be using both habitats.\n    </para>"
#> 
#> $abstract$section[[6]]
#> [1] "<title>Snorkel Index Data</title>\n<para>\n      This dataset covers salmonid observations conducted in the\n      Sacramento River and focused on assessing effectiveness of\n      salmonid habitat improvement projects.  Surveys are conducted\n      roughly every other week and include pre-project sites,\n      constructed habitat project sites, and control sites where no\n      treatment is planned.  Sites are snorkeled by a crew of two to\n      three people and all salmonids are counted by size category within\n      pre-set survey reaches.  Snorkelers record data on dive slates and\n      then transcribe to paper data sheets and computer files.  Annual\n      reports summarize the survey findings.\n    </para>"
#> 
#> 
#> $abstract$para
#> list()
#> 
#>