Add Coverage

Adds the coverage metadata elements to a dataset list according to EML standards.

add_coverage(parent_element, coverage_metadata, taxonomic_metadata = NULL)

Arguments

parent_element

A list representing the EML project or dataset.

coverage_metadata

a list or datatable of coverage metadata see create_coverage

taxonomic_metadata

a list or datatable of taxonomic metadata see create_taxonomic_coverage

Value

The dataset list or project with coverage information appended.

Examples

coverage <- list(geographic_description = "Description",
             west_bounding_coordinate = "-160.594000", 
             east_bounding_coordinate = "-134.104800",
             north_bounding_coordinate = "71.238300", 
             south_bounding_coordinate = "67.865000",
             begin_date = "1980-01-01", 
             end_date = "2010-12-31")
if (FALSE) {             
dataset <- list() %>% 
  add_coverage(coverage)
dataset
}