Add Data Table

Adds the data table elements to a dataset list according to EML standards.

add_datatable(parent_element, datatable_metadata)

Arguments

parent_element
A list representing the EML project or dataset.

datatable_metadata
A named list or dataframe containing datatable metadata elements (filepath, attribute_info, datatable_description, datatable_url): see create_datatable

Value

The dataset list or project with datatable information appended.

Examples

datatable_metadata <- dplyr::tibble(filepath = c(system.file("extdata", "Banet-Example", "data", "enclosure-study-growth-rate-data.csv", package = "EMLaide", mustWork = TRUE), 
                                                             system.file("extdata", "Banet-Example", "data", "enclosure-study-gut-contents-data.csv", package = "EMLaide", mustWork = TRUE)), 
                                     attribute_info = c(system.file("extdata", "Banet-Example", "metadata", "enclosure-study-growth-rates-metadata.xlsx", package = "EMLaide", mustWork = TRUE), 
                                     system.file("extdata", "Banet-Example", "metadata", "enclosure-study-gut-contents-metadata.xlsx", package = "EMLaide", mustWork = TRUE)),
                                     datatable_description = c("Growth Rates - Enclosure Study","Gut Contents - Enclosure Study"))
 
dataset <- list() %>%
    add_datatable(datatable_metadata)

#> [1] "We identified the following custom unit: fishPerEnclosure , please make sure to add information on this custom unit in additional metadata information:"
#> [1] "We identified the following custom unit: fishPerEnclosure , please make sure to add information on this custom unit in additional metadata information:"
#> [1] "We identified the following custom unit: fishPerEnclosure , please make sure to add information on this custom unit in additional metadata information:"
#> [1] "We identified the following custom unit: thermal unit , please make sure to add information on this custom unit in additional metadata information:"
#> [1] "We identified the following custom unit: day , please make sure to add information on this custom unit in additional metadata information:"

dataset

#> $dataTable
#> $dataTable[[1]]
#> $dataTable[[1]]$entityName
#> [1] "enclosure-study-growth-rate-data.csv"
#> 
#> $dataTable[[1]]$entityDescription
#> [1] "Growth Rates - Enclosure Study"
#> 
#> $dataTable[[1]]$physical
#> $dataTable[[1]]$physical$objectName
#> [1] "enclosure-study-growth-rate-data.csv"
#> 
#> $dataTable[[1]]$physical$size
#> $dataTable[[1]]$physical$size$unit
#> [1] "bytes"
#> 
#> $dataTable[[1]]$physical$size$size
#> [1] "32636"
#> 
#> 
#> $dataTable[[1]]$physical$authentication
#> $dataTable[[1]]$physical$authentication$method
#> [1] "MD5"
#> 
#> $dataTable[[1]]$physical$authentication$authentication
#> [1] "44b508c45e3635e8d65a719631530a75"
#> 
#> 
#> $dataTable[[1]]$physical$dataFormat
#> $dataTable[[1]]$physical$dataFormat$textFormat
#> $dataTable[[1]]$physical$dataFormat$textFormat$numHeaderLines
#> [1] "1"
#> 
#> $dataTable[[1]]$physical$dataFormat$textFormat$recordDelimiter
#> [1] "\\r\\n"
#> 
#> $dataTable[[1]]$physical$dataFormat$textFormat$attributeOrientation
#> [1] "column"
#> 
#> $dataTable[[1]]$physical$dataFormat$textFormat$simpleDelimited
#> $dataTable[[1]]$physical$dataFormat$textFormat$simpleDelimited$fieldDelimiter
#> [1] ","
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$physical$distribution
#> $dataTable[[1]]$physical$distribution$online
#> $dataTable[[1]]$physical$distribution$online$url
#> $dataTable[[1]]$physical$distribution$online$url$url
#> NULL
#> 
#> $dataTable[[1]]$physical$distribution$online$url$`function`
#> [1] "download"
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList
#> $dataTable[[1]]$attributeList$attribute
#> $dataTable[[1]]$attributeList$attribute[[1]]
#> $dataTable[[1]]$attributeList$attribute[[1]]$attributeName
#> [1] "enclosure"
#> 
#> $dataTable[[1]]$attributeList$attribute[[1]]$attributeDefinition
#> [1] "Enclosure ID. Enclosures were number 1-6 in each site for identification."
#> 
#> $dataTable[[1]]$attributeList$attribute[[1]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[1]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[1]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[1]]$measurementScale$nominal
#> $dataTable[[1]]$attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain
#> $dataTable[[1]]$attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $dataTable[[1]]$attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "Enclosure ID. Enclosures were number 1-6 in each site for identification."
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[2]]
#> $dataTable[[1]]$attributeList$attribute[[2]]$attributeName
#> [1] "fish_code"
#> 
#> $dataTable[[1]]$attributeList$attribute[[2]]$attributeDefinition
#> [1] "Unique identifying number for fish within an enclosure. Fish were numbered 1-20 within each enclosure."
#> 
#> $dataTable[[1]]$attributeList$attribute[[2]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[2]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[2]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[2]]$measurementScale$nominal
#> $dataTable[[1]]$attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain
#> $dataTable[[1]]$attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $dataTable[[1]]$attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "Unique identifying number for fish within an enclosure. Fish were numbered 1-20 within each enclosure."
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]
#> $dataTable[[1]]$attributeList$attribute[[3]]$attributeName
#> [1] "length_begin"
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$attributeDefinition
#> [1] "fish fork length at the start of the study"
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$unit$standardUnit
#> [1] "millimeter"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 48
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 78
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]
#> $dataTable[[1]]$attributeList$attribute[[4]]$attributeName
#> [1] "length_middle"
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$attributeDefinition
#> [1] "fish fork length at the mid-point of the study"
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$unit$standardUnit
#> [1] "millimeter"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 68
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 84
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]
#> $dataTable[[1]]$attributeList$attribute[[5]]$attributeName
#> [1] "length_end"
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$attributeDefinition
#> [1] "fish fork length at the end of the study"
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$unit$standardUnit
#> [1] "millimeter"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 65
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 95
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]
#> $dataTable[[1]]$attributeList$attribute[[6]]$attributeName
#> [1] "condition_beginning"
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$attributeDefinition
#> [1] "Fulton's condition factor at the beginning of the study"
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0.3350023
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 2.700319
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]
#> $dataTable[[1]]$attributeList$attribute[[7]]$attributeName
#> [1] "condition_ending"
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$attributeDefinition
#> [1] "Fulton's condition factor at the end of the study"
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0.7327568
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1.590976
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]
#> $dataTable[[1]]$attributeList$attribute[[8]]$attributeName
#> [1] "density_start"
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$attributeDefinition
#> [1] "density at the start of the study, represented as the number of fish in the enclosure"
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$unit$customUnit
#> [1] "fishPerEnclosure"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 18
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 20
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]
#> $dataTable[[1]]$attributeList$attribute[[9]]$attributeName
#> [1] "density_middle"
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$attributeDefinition
#> [1] "density at the mid-point of the study, represented as the number of fish in the enclosure"
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$unit$customUnit
#> [1] "fishPerEnclosure"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 11
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 20
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]
#> $dataTable[[1]]$attributeList$attribute[[10]]$attributeName
#> [1] "density_end"
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$attributeDefinition
#> [1] "density at the end of the study, represented as the number of fish in the enclosure"
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$unit$customUnit
#> [1] "fishPerEnclosure"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 5
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 20
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]
#> $dataTable[[1]]$attributeList$attribute[[11]]$attributeName
#> [1] "atu"
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$attributeDefinition
#> [1] "acumulated thermal units experienced by the fish over the course of the study"
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$unit$customUnit
#> [1] "thermal unit"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 593.193
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 628.55
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[12]]
#> $dataTable[[1]]$attributeList$attribute[[12]]$attributeName
#> [1] "habitat"
#> 
#> $dataTable[[1]]$attributeList$attribute[[12]]$attributeDefinition
#> [1] "The habitat type (mainstem/sidechannel) and treatment (control/restored) for each study site"
#> 
#> $dataTable[[1]]$attributeList$attribute[[12]]$storageType
#> [1] "string"
#> 
#> $dataTable[[1]]$attributeList$attribute[[12]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[12]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[12]]$measurementScale$nominal
#> $dataTable[[1]]$attributeList$attribute[[12]]$measurementScale$nominal$nonNumericDomain
#> $dataTable[[1]]$attributeList$attribute[[12]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $dataTable[[1]]$attributeList$attribute[[12]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "The habitat type (mainstem/sidechannel) and treatment (control/restored) for each study site"
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[13]]
#> $dataTable[[1]]$attributeList$attribute[[13]]$attributeName
#> [1] "site_id"
#> 
#> $dataTable[[1]]$attributeList$attribute[[13]]$attributeDefinition
#> [1] "site name for each study site"
#> 
#> $dataTable[[1]]$attributeList$attribute[[13]]$storageType
#> [1] "string"
#> 
#> $dataTable[[1]]$attributeList$attribute[[13]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[13]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[13]]$measurementScale$nominal
#> $dataTable[[1]]$attributeList$attribute[[13]]$measurementScale$nominal$nonNumericDomain
#> $dataTable[[1]]$attributeList$attribute[[13]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $dataTable[[1]]$attributeList$attribute[[13]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "site name for each study site"
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]
#> $dataTable[[1]]$attributeList$attribute[[14]]$attributeName
#> [1] "days"
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$attributeDefinition
#> [1] "number of days fish were in enclosure"
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$unit
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$unit$customUnit
#> [1] "day"
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 43
#> 
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[1]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 48
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[1]]$numberOfRecords
#> [1] 353
#> 
#> 
#> $dataTable[[2]]
#> $dataTable[[2]]$entityName
#> [1] "enclosure-study-gut-contents-data.csv"
#> 
#> $dataTable[[2]]$entityDescription
#> [1] "Gut Contents - Enclosure Study"
#> 
#> $dataTable[[2]]$physical
#> $dataTable[[2]]$physical$objectName
#> [1] "enclosure-study-gut-contents-data.csv"
#> 
#> $dataTable[[2]]$physical$size
#> $dataTable[[2]]$physical$size$unit
#> [1] "bytes"
#> 
#> $dataTable[[2]]$physical$size$size
#> [1] "4899"
#> 
#> 
#> $dataTable[[2]]$physical$authentication
#> $dataTable[[2]]$physical$authentication$method
#> [1] "MD5"
#> 
#> $dataTable[[2]]$physical$authentication$authentication
#> [1] "b72c72b6b1db5b82c163379f8ee5f9a3"
#> 
#> 
#> $dataTable[[2]]$physical$dataFormat
#> $dataTable[[2]]$physical$dataFormat$textFormat
#> $dataTable[[2]]$physical$dataFormat$textFormat$numHeaderLines
#> [1] "1"
#> 
#> $dataTable[[2]]$physical$dataFormat$textFormat$recordDelimiter
#> [1] "\\r\\n"
#> 
#> $dataTable[[2]]$physical$dataFormat$textFormat$attributeOrientation
#> [1] "column"
#> 
#> $dataTable[[2]]$physical$dataFormat$textFormat$simpleDelimited
#> $dataTable[[2]]$physical$dataFormat$textFormat$simpleDelimited$fieldDelimiter
#> [1] ","
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$physical$distribution
#> $dataTable[[2]]$physical$distribution$online
#> $dataTable[[2]]$physical$distribution$online$url
#> $dataTable[[2]]$physical$distribution$online$url$url
#> NULL
#> 
#> $dataTable[[2]]$physical$distribution$online$url$`function`
#> [1] "download"
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList
#> $dataTable[[2]]$attributeList$attribute
#> $dataTable[[2]]$attributeList$attribute[[1]]
#> $dataTable[[2]]$attributeList$attribute[[1]]$attributeName
#> [1] "habitat"
#> 
#> $dataTable[[2]]$attributeList$attribute[[1]]$attributeDefinition
#> [1] "The habitat type (mainstem/sidechannel) and treatment (control/restored) for each study site"
#> 
#> $dataTable[[2]]$attributeList$attribute[[1]]$storageType
#> [1] "string"
#> 
#> $dataTable[[2]]$attributeList$attribute[[1]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[1]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[1]]$measurementScale$nominal
#> $dataTable[[2]]$attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain
#> $dataTable[[2]]$attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $dataTable[[2]]$attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "The habitat type (mainstem/sidechannel) and treatment (control/restored) for each study site"
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[2]]
#> $dataTable[[2]]$attributeList$attribute[[2]]$attributeName
#> [1] "site_id"
#> 
#> $dataTable[[2]]$attributeList$attribute[[2]]$attributeDefinition
#> [1] "site name for each study site"
#> 
#> $dataTable[[2]]$attributeList$attribute[[2]]$storageType
#> [1] "string"
#> 
#> $dataTable[[2]]$attributeList$attribute[[2]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[2]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[2]]$measurementScale$nominal
#> $dataTable[[2]]$attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain
#> $dataTable[[2]]$attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $dataTable[[2]]$attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "site name for each study site"
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[3]]
#> $dataTable[[2]]$attributeList$attribute[[3]]$attributeName
#> [1] "enclosure_number"
#> 
#> $dataTable[[2]]$attributeList$attribute[[3]]$attributeDefinition
#> [1] "Enclosure ID. Enclosures uniquely identified."
#> 
#> $dataTable[[2]]$attributeList$attribute[[3]]$storageType
#> [1] "string"
#> 
#> $dataTable[[2]]$attributeList$attribute[[3]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[3]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[3]]$measurementScale$nominal
#> $dataTable[[2]]$attributeList$attribute[[3]]$measurementScale$nominal$nonNumericDomain
#> $dataTable[[2]]$attributeList$attribute[[3]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $dataTable[[2]]$attributeList$attribute[[3]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "Enclosure ID. Enclosures uniquely identified."
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]
#> $dataTable[[2]]$attributeList$attribute[[4]]$attributeName
#> [1] "diptera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$attributeDefinition
#> [1] "Number of diptera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]
#> $dataTable[[2]]$attributeList$attribute[[5]]$attributeName
#> [1] "exuvium"
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$attributeDefinition
#> [1] "Number of exuvium found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]
#> $dataTable[[2]]$attributeList$attribute[[6]]$attributeName
#> [1] "coleoptera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$attributeDefinition
#> [1] "Number of coleoptera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]
#> $dataTable[[2]]$attributeList$attribute[[7]]$attributeName
#> [1] "hemiptera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$attributeDefinition
#> [1] "Number of hemiptera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]
#> $dataTable[[2]]$attributeList$attribute[[8]]$attributeName
#> [1] "odonata"
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$attributeDefinition
#> [1] "Number of odonata found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]
#> $dataTable[[2]]$attributeList$attribute[[9]]$attributeName
#> [1] "trichoptera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$attributeDefinition
#> [1] "Number of trichoptera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]
#> $dataTable[[2]]$attributeList$attribute[[10]]$attributeName
#> [1] "ephemeroptera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$attributeDefinition
#> [1] "Number of ephemeroptera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]
#> $dataTable[[2]]$attributeList$attribute[[11]]$attributeName
#> [1] "plecoptera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$attributeDefinition
#> [1] "Number of plecoptera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]
#> $dataTable[[2]]$attributeList$attribute[[12]]$attributeName
#> [1] "hymenoptera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$attributeDefinition
#> [1] "Number of hymenoptera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[12]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]
#> $dataTable[[2]]$attributeList$attribute[[13]]$attributeName
#> [1] "ostracoda"
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$attributeDefinition
#> [1] "Number of ostracoda found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[13]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]
#> $dataTable[[2]]$attributeList$attribute[[14]]$attributeName
#> [1] "cladocera"
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$attributeDefinition
#> [1] "Number of cladocera found in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]
#> $dataTable[[2]]$attributeList$attribute[[15]]$attributeName
#> [1] "amphipoda"
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$attributeDefinition
#> [1] "Number of amphipodafound in the gut contents"
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[15]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]
#> $dataTable[[2]]$attributeList$attribute[[16]]$attributeName
#> [1] "Sum"
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$attributeDefinition
#> [1] "Total food items found in gut"
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$storageType
#> [1] "dbl"
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$attributeLabel
#> [1] NA
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$unit
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$bounds
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0
#> 
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $dataTable[[2]]$attributeList$attribute[[16]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1e+08
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $dataTable[[2]]$numberOfRecords
#> [1] 72
#> 
#> 
#>