Add Methods

Adds the methods metadata elements according to EML standards.

add_method(parent_element, methods_file)

Arguments

parent_element
A list representing the EML project or dataset.

methods_file
A file containing methods information (word or markdown): see create_methods

Value

The dataset list or project with methods information appended.

Examples

methods_docx <- methods_docx <- system.file("extdata", "Banet-Example", "metadata", "methods.docx", 
                                             package = "EMLaide", mustWork = TRUE)
dataset <- list() %>%
    add_method(methods_docx)
dataset

#> $methods
#> $methods$sampling
#> NULL
#> 
#> $methods$methodStep
#> $methods$methodStep$instrumentation
#> character(0)
#> 
#> $methods$methodStep$software
#> NULL
#> 
#> $methods$methodStep$description
#> $methods$methodStep$description$section
#> $methods$methodStep$description$section[[1]]
#> [1] "<title>Enclosure Study: Growth Rate Methods.</title>\n<para>\n      Data was collected following methods described at the following\n      link:\n    </para>\n<para>\n  <link xlink:href=\"https://www.sacramentoriver.org/forum/publications/side_channels/Monitoring%20Plan%20(final)%20170831.pdf\">\n    <emphasis role=\"underline\">https://www.sacramentoriver.org/forum/publications/side_channels/Monitoring%20Plan%20(final)%20170831.pdf</emphasis>\n  </link>\n</para>\n<para>\n      Monitoring efforts will generate data for the presented response\n      variables for the purpose of testing the hypotheses presented in\n      table 1 of the aforementioned document. Specific response variable\n      metrics are identified while methods reference field protocols\n      that are documented in Appendix A (Table A-1).\n    </para>\n<para>\n      Note that the above referenced protocol includes additional types\n      of data collection. Enclosure study methods are identified and\n      clearly labeled throughout the document.\n    </para>\n<para>\n      Caveats to using this data: Some enclosures lost fish due to\n      mortality or escape, and others did not. This resulted in\n      different densities between habitat types, and confounded habitat\n      type with density dependent effects on growth. Logistical\n      constraints prevented removal of all enclosures on the same day,\n      so the total time from start to finish was not identical between\n      habitat types.\n    </para>"
#> 
#> 
#> $methods$methodStep$description$para
#> list()
#> 
#> 
#> 
#>