Create Data Table

Creates the data table elements according to EML standards.

create_datatable(
  filepath,
  attribute_info,
  datatable_description,
  datatable_url = NULL,
  dataset_methods = NULL,
  additional_info = NULL
)

Arguments

attribute_info
File path to the attribute sheet that contains metadata describing all attributes for your datatable.

datatable_description
Short description of contents of datatable you are adding

datatable_url
A URL that links to the datatable you are adding (ex: dataset on AWS public bucket or https://raw.githubusercontent.com/)

dataset_methods
Optional metadata describing methods if there are methods that are specific to a datatable

additional_info
Option additional metadata info if there is additional metadata information that is important to the datatable

filename
Name of the file you would like added to the datatable element

Value

A data table element listed in a format to be added to a dataset element and turned into EML

Details

Create Data Table

Examples

filepath =  system.file("extdata", "Banet-Example", "data", "enclosure-study-growth-rate-data.csv", package = "EMLaide", mustWork = TRUE)
attribute_info = system.file("extdata", "Banet-Example", "metadata", "enclosure-study-growth-rates-metadata.xlsx", package = "EMLaide", mustWork = TRUE)
datatable_description = "Growth Rates - Enclosure Study"

create_datatable(filepath, attribute_info, datatable_description)

#> Warning: Please add text pattern if applicable.

#> Warning: Please add text pattern if applicable.

#> [1] "We identified the following custom unit: fishPerEnclosure , please make sure to add information on this custom unit in additional metadata information:"
#> [1] "We identified the following custom unit: fishPerEnclosure , please make sure to add information on this custom unit in additional metadata information:"
#> [1] "We identified the following custom unit: fishPerEnclosure , please make sure to add information on this custom unit in additional metadata information:"
#> [1] "We identified the following custom unit: thermal unit , please make sure to add information on this custom unit in additional metadata information:"

#> Warning: Please add text pattern if applicable.

#> Warning: Please add text pattern if applicable.

#> [1] "We identified the following custom unit: day , please make sure to add information on this custom unit in additional metadata information:"

#> $entityName
#> [1] "enclosure-study-growth-rate-data.csv"
#> 
#> $entityDescription
#> [1] "Growth Rates - Enclosure Study"
#> 
#> $physical
#> $physical$objectName
#> [1] "enclosure-study-growth-rate-data.csv"
#> 
#> $physical$size
#> $physical$size$unit
#> [1] "bytes"
#> 
#> $physical$size$size
#> [1] "32636"
#> 
#> 
#> $physical$authentication
#> $physical$authentication$method
#> [1] "MD5"
#> 
#> $physical$authentication$authentication
#> [1] "44b508c45e3635e8d65a719631530a75"
#> 
#> 
#> $physical$dataFormat
#> $physical$dataFormat$textFormat
#> $physical$dataFormat$textFormat$numHeaderLines
#> [1] "1"
#> 
#> $physical$dataFormat$textFormat$recordDelimiter
#> [1] "\\r\\n"
#> 
#> $physical$dataFormat$textFormat$attributeOrientation
#> [1] "column"
#> 
#> $physical$dataFormat$textFormat$simpleDelimited
#> $physical$dataFormat$textFormat$simpleDelimited$fieldDelimiter
#> [1] ","
#> 
#> 
#> 
#> 
#> $physical$distribution
#> $physical$distribution$online
#> $physical$distribution$online$url
#> $physical$distribution$online$url$url
#> NULL
#> 
#> $physical$distribution$online$url$`function`
#> [1] "download"
#> 
#> 
#> 
#> 
#> 
#> $attributeList
#> $attributeList$attribute
#> $attributeList$attribute[[1]]
#> $attributeList$attribute[[1]]$attributeName
#> [1] "enclosure"
#> 
#> $attributeList$attribute[[1]]$attributeDefinition
#> [1] "Enclosure ID. Enclosures were number 1-6 in each site for identification."
#> 
#> $attributeList$attribute[[1]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[1]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[1]]$measurementScale
#> $attributeList$attribute[[1]]$measurementScale$nominal
#> $attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain
#> $attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $attributeList$attribute[[1]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "Enclosure ID. Enclosures were number 1-6 in each site for identification."
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[2]]
#> $attributeList$attribute[[2]]$attributeName
#> [1] "fish_code"
#> 
#> $attributeList$attribute[[2]]$attributeDefinition
#> [1] "Unique identifying number for fish within an enclosure. Fish were numbered 1-20 within each enclosure."
#> 
#> $attributeList$attribute[[2]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[2]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[2]]$measurementScale
#> $attributeList$attribute[[2]]$measurementScale$nominal
#> $attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain
#> $attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $attributeList$attribute[[2]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "Unique identifying number for fish within an enclosure. Fish were numbered 1-20 within each enclosure."
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[3]]
#> $attributeList$attribute[[3]]$attributeName
#> [1] "length_begin"
#> 
#> $attributeList$attribute[[3]]$attributeDefinition
#> [1] "fish fork length at the start of the study"
#> 
#> $attributeList$attribute[[3]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[3]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[3]]$measurementScale
#> $attributeList$attribute[[3]]$measurementScale$ratio
#> $attributeList$attribute[[3]]$measurementScale$ratio$unit
#> $attributeList$attribute[[3]]$measurementScale$ratio$unit$standardUnit
#> [1] "millimeter"
#> 
#> 
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 48
#> 
#> 
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[3]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 78
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[4]]
#> $attributeList$attribute[[4]]$attributeName
#> [1] "length_middle"
#> 
#> $attributeList$attribute[[4]]$attributeDefinition
#> [1] "fish fork length at the mid-point of the study"
#> 
#> $attributeList$attribute[[4]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[4]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[4]]$measurementScale
#> $attributeList$attribute[[4]]$measurementScale$ratio
#> $attributeList$attribute[[4]]$measurementScale$ratio$unit
#> $attributeList$attribute[[4]]$measurementScale$ratio$unit$standardUnit
#> [1] "millimeter"
#> 
#> 
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 68
#> 
#> 
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[4]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 84
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[5]]
#> $attributeList$attribute[[5]]$attributeName
#> [1] "length_end"
#> 
#> $attributeList$attribute[[5]]$attributeDefinition
#> [1] "fish fork length at the end of the study"
#> 
#> $attributeList$attribute[[5]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[5]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[5]]$measurementScale
#> $attributeList$attribute[[5]]$measurementScale$ratio
#> $attributeList$attribute[[5]]$measurementScale$ratio$unit
#> $attributeList$attribute[[5]]$measurementScale$ratio$unit$standardUnit
#> [1] "millimeter"
#> 
#> 
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 65
#> 
#> 
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[5]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 95
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[6]]
#> $attributeList$attribute[[6]]$attributeName
#> [1] "condition_beginning"
#> 
#> $attributeList$attribute[[6]]$attributeDefinition
#> [1] "Fulton's condition factor at the beginning of the study"
#> 
#> $attributeList$attribute[[6]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[6]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[6]]$measurementScale
#> $attributeList$attribute[[6]]$measurementScale$ratio
#> $attributeList$attribute[[6]]$measurementScale$ratio$unit
#> $attributeList$attribute[[6]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0.3350023
#> 
#> 
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[6]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 2.700319
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[7]]
#> $attributeList$attribute[[7]]$attributeName
#> [1] "condition_ending"
#> 
#> $attributeList$attribute[[7]]$attributeDefinition
#> [1] "Fulton's condition factor at the end of the study"
#> 
#> $attributeList$attribute[[7]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[7]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[7]]$measurementScale
#> $attributeList$attribute[[7]]$measurementScale$ratio
#> $attributeList$attribute[[7]]$measurementScale$ratio$unit
#> $attributeList$attribute[[7]]$measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 0.7327568
#> 
#> 
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[7]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 1.590976
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[8]]
#> $attributeList$attribute[[8]]$attributeName
#> [1] "density_start"
#> 
#> $attributeList$attribute[[8]]$attributeDefinition
#> [1] "density at the start of the study, represented as the number of fish in the enclosure"
#> 
#> $attributeList$attribute[[8]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[8]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[8]]$measurementScale
#> $attributeList$attribute[[8]]$measurementScale$ratio
#> $attributeList$attribute[[8]]$measurementScale$ratio$unit
#> $attributeList$attribute[[8]]$measurementScale$ratio$unit$customUnit
#> [1] "fishPerEnclosure"
#> 
#> 
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 18
#> 
#> 
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[8]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 20
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[9]]
#> $attributeList$attribute[[9]]$attributeName
#> [1] "density_middle"
#> 
#> $attributeList$attribute[[9]]$attributeDefinition
#> [1] "density at the mid-point of the study, represented as the number of fish in the enclosure"
#> 
#> $attributeList$attribute[[9]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[9]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[9]]$measurementScale
#> $attributeList$attribute[[9]]$measurementScale$ratio
#> $attributeList$attribute[[9]]$measurementScale$ratio$unit
#> $attributeList$attribute[[9]]$measurementScale$ratio$unit$customUnit
#> [1] "fishPerEnclosure"
#> 
#> 
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 11
#> 
#> 
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[9]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 20
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[10]]
#> $attributeList$attribute[[10]]$attributeName
#> [1] "density_end"
#> 
#> $attributeList$attribute[[10]]$attributeDefinition
#> [1] "density at the end of the study, represented as the number of fish in the enclosure"
#> 
#> $attributeList$attribute[[10]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[10]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[10]]$measurementScale
#> $attributeList$attribute[[10]]$measurementScale$ratio
#> $attributeList$attribute[[10]]$measurementScale$ratio$unit
#> $attributeList$attribute[[10]]$measurementScale$ratio$unit$customUnit
#> [1] "fishPerEnclosure"
#> 
#> 
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 5
#> 
#> 
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[10]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 20
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[11]]
#> $attributeList$attribute[[11]]$attributeName
#> [1] "atu"
#> 
#> $attributeList$attribute[[11]]$attributeDefinition
#> [1] "acumulated thermal units experienced by the fish over the course of the study"
#> 
#> $attributeList$attribute[[11]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[11]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[11]]$measurementScale
#> $attributeList$attribute[[11]]$measurementScale$ratio
#> $attributeList$attribute[[11]]$measurementScale$ratio$unit
#> $attributeList$attribute[[11]]$measurementScale$ratio$unit$customUnit
#> [1] "thermal unit"
#> 
#> 
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 593.193
#> 
#> 
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[11]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 628.55
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[12]]
#> $attributeList$attribute[[12]]$attributeName
#> [1] "habitat"
#> 
#> $attributeList$attribute[[12]]$attributeDefinition
#> [1] "The habitat type (mainstem/sidechannel) and treatment (control/restored) for each study site"
#> 
#> $attributeList$attribute[[12]]$storageType
#> [1] "string"
#> 
#> $attributeList$attribute[[12]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[12]]$measurementScale
#> $attributeList$attribute[[12]]$measurementScale$nominal
#> $attributeList$attribute[[12]]$measurementScale$nominal$nonNumericDomain
#> $attributeList$attribute[[12]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $attributeList$attribute[[12]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "The habitat type (mainstem/sidechannel) and treatment (control/restored) for each study site"
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[13]]
#> $attributeList$attribute[[13]]$attributeName
#> [1] "site_id"
#> 
#> $attributeList$attribute[[13]]$attributeDefinition
#> [1] "site name for each study site"
#> 
#> $attributeList$attribute[[13]]$storageType
#> [1] "string"
#> 
#> $attributeList$attribute[[13]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[13]]$measurementScale
#> $attributeList$attribute[[13]]$measurementScale$nominal
#> $attributeList$attribute[[13]]$measurementScale$nominal$nonNumericDomain
#> $attributeList$attribute[[13]]$measurementScale$nominal$nonNumericDomain$textDomain
#> $attributeList$attribute[[13]]$measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "site name for each study site"
#> 
#> 
#> 
#> 
#> 
#> 
#> $attributeList$attribute[[14]]
#> $attributeList$attribute[[14]]$attributeName
#> [1] "days"
#> 
#> $attributeList$attribute[[14]]$attributeDefinition
#> [1] "number of days fish were in enclosure"
#> 
#> $attributeList$attribute[[14]]$storageType
#> [1] "dbl"
#> 
#> $attributeList$attribute[[14]]$attributeLabel
#> [1] NA
#> 
#> $attributeList$attribute[[14]]$measurementScale
#> $attributeList$attribute[[14]]$measurementScale$ratio
#> $attributeList$attribute[[14]]$measurementScale$ratio$unit
#> $attributeList$attribute[[14]]$measurementScale$ratio$unit$customUnit
#> [1] "day"
#> 
#> 
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$numberType
#> [1] "whole"
#> 
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$minimum$minimum
#> [1] 43
#> 
#> 
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $attributeList$attribute[[14]]$measurementScale$ratio$numericDomain$bounds$maximum$maximum
#> [1] 48
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $numberOfRecords
#> [1] 353
#>