Create an attribute in dateTime Measurement Scale

Create an attribute in dateTime Measurement Scale

create_datetime(date_time_format, date_time_precision, minimum, maximum)

Arguments

date_time_format
ISO 8601 format should be used.

date_time_precision
To what level of time your attribute is recorded.

minimum
The earliest dateTime recorded. Values can be larger than or equal to this number.

maximum
The latest dateTime recorded. Values can be less than or equal to this number.

Examples

create_attribute(attribute_name = "Yrs",
              attribute_definition = "Calendar year of the observation from years 1990 - 2010.",
              storage_type = EMLaide::storage_type$integer,
              measurement_scale = EMLaide::measurement_scale$dateTime,
              attribute_label = "Years",
              date_time_format = "YYYY", 
              date_time_precision = "1", 
              minimum = "1993", 
              maximum = "2003")

#> $attributeName
#> [1] "Yrs"
#> 
#> $attributeDefinition
#> [1] "Calendar year of the observation from years 1990 - 2010."
#> 
#> $storageType
#> NULL
#> 
#> $attributeLabel
#> [1] "Years"
#> 
#> $measurementScale
#> $measurementScale$dateTime
#> $measurementScale$dateTime$formatString
#> [1] "YYYY"
#> 
#> $measurementScale$dateTime$dateTimePrecision
#> [1] "1"
#> 
#> $measurementScale$dateTime$dateTimeDomain
#> $measurementScale$dateTime$dateTimeDomain$bounds
#> $measurementScale$dateTime$dateTimeDomain$bounds$minimum
#> $measurementScale$dateTime$dateTimeDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $measurementScale$dateTime$dateTimeDomain$bounds$minimum$minimum
#> [1] "1993"
#> 
#> 
#> $measurementScale$dateTime$dateTimeDomain$bounds$maximum
#> $measurementScale$dateTime$dateTimeDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $measurementScale$dateTime$dateTimeDomain$bounds$maximum$maximum
#> [1] "2003"
#> 
#> 
#> 
#> 
#> 
#>