Create project element

This function creates a project element from metadata inputs using create_person and create_funding.

create_project(project_title, project_lead, funding_metadata)

Arguments

project_title

A project title as a string

project_lead

A named list or datatable describing the project lead for the project. See structure by looking at required inputs of create_person

funding_metadata

A named list or datatable describing the project funding. See structure by looking at required inputs of create_funding

Value

This function returns project element containing all project information required for an EML document.

Examples

project_title <- "Salmonid Habitat Monitoring"
project_lead <- list(first_name = "Stacy", last_name = "Banet", email = "Stacy@aol.com", 
                                   role = "creator", organization = "USBR")
funding_metadata <- list(funder_name = "USBR", funder_identifier = NA, award_number = "R14AC00096", 
                         award_title = "Salmonid Spawning and Rearing Habitat Restoration in the Sacramento River", 
                         award_url = NA)

create_project(project_title, project_lead, funding_metadata)
#> $title
#> [1] "Salmonid Habitat Monitoring"
#> 
#> $personnel
#> $personnel$individualName
#> $personnel$individualName$givenName
#> [1] "Stacy"
#> 
#> $personnel$individualName$surName
#> [1] "Banet"
#> 
#> 
#> $personnel$electronicMailAddress
#> [1] "Stacy@aol.com"
#> 
#> $personnel$organizationName
#> [1] "USBR"
#> 
#> $personnel$role
#> [1] "Project Lead"
#> 
#> 
#> $award
#> $award[[1]]
#> $award[[1]]$funderName
#> [1] "United States Bureau of Reclamation"
#> 
#> $award[[1]]$funderIdentifier
#> [1] "https://www.wikidata.org/wiki/Q1010548"
#> 
#> $award[[1]]$title
#> [1] "Salmonid Spawning and Rearing Habitat Restoration in the Sacramento River"
#> 
#> $award[[1]]$awardNumber
#> [1] "R14AC00096"
#> 
#> $award[[1]]$awardUrl
#> [1] NA
#> 
#> 
#>