Create Taxonomic Coverage

Create the taxonomic coverage information of a dataset based off of EML standards. The addition of taxonomic coverage is (Optional), however defaults of chinook, sturgeon, smelt, and steelhead are provided. While single or multiple taxonomies can be applied, the full set of information must be provided if chosen to be included.

create_taxonomic_coverage(
  CVPIA_common_species = NULL,
  kingdom = "kingdom",
  kingdom_value = NULL,
  phylum = "phylum",
  phylum_value = NULL,
  class = "class",
  class_value = NULL,
  order = "order",
  order_value = NULL,
  family = "family",
  family_value = NULL,
  genus = "genus",
  genus_value = NULL,
  species = "species",
  species_value = NULL,
  common_name = NULL,
  taxon_id = NULL
)

Arguments

CVPIA_common_species

Use one of the following from the helper data CVPIA_common_species: "chinook", "delta_smelt", "white_sturgeon", "green_sturgeon", or "steelhead" to get pre-selected information from ITIS.

kingdom

Kingdom level present.

kingdom_value

The kingdom name.

phylum

Phylum level present.

phylum_value

The phylum name.

class

Class level present.

class_value

The class level name.

order

Order level present.

order_value

The order level name.

family

Family level present.

family_value

The family level name.

genus

Genus level present.

genus_value

The genus level name.

species

Species level present.

species_value

The species level name.

common_name

The common name of the organism.

taxon_id

(Optional). The taxonomic serial number provided by ITIS.

Value

Taxonomic coverage information. The function should be assigned to the name taxonomic_coverage to append it to the dataset or project. Example of how to incorporate it into the add_coverage function is seen below.

CVPIA Common Species

The following frequently cited species are available for convenience:

EMLaide::CVPIA_common_species$chinook - Oncorhynchus tshawytscha (ITIS: 161980)

EMLaide::CVPIA_common_species$steelhead - Oncorhynchus mykiss (ITIS: 161989)

EMLaide::CVPIA_common_species$delta_smelt - Hypomesus transpacificus (ITIS: 162032)

EMLaide::CVPIA_common_species$white_sturgeon - Acipenser transmontanus (ITIS:161068)

EMLaide::CVPIA_common_species$green_sturgeon - Acipenser medirostris (ITIS: 161067)

For further taxonomic coverage (i.e subkingdom, infrakingdom, etc.) on any of these species, you can visit: ITIS's webpage for full coverage information.

Examples

if (FALSE) taxonomic_coverage <- create_taxonomic_coverage(CVPIA_common_species = "chinook")

if (FALSE) taxonomic_coverage <- create_taxonomic_coverage(CVPIA_common_species = "delta_smelt")

if (FALSE) taxonomic_coverage <- create_taxonomic_coverage(CVPIA_common_species = "white_sturgeon")

if (FALSE) taxonomic_coverage <- create_taxonomic_coverage(CVPIA_common_species = "green_sturgeon")

if (FALSE) taxonomic_coverage <- create_taxonomic_coverage(CVPIA_common_species = "steelhead")

if (FALSE) taxonomic_coverage <- create_taxonomic_coverage(kingdom_value = "Animalia",
                                                         phylum_value = "Chordata",
                                                         class_value = "Mammalia",
                                                         order_value = "Carnivora",
                                                         family_value = "Felidae",
                                                         genus_value = "Panthera", 
                                                         species_value = "Panthera Leo",
                                                         common_name = "Lion", 
                                                         taxon_id = "183803")
                                            
# To append this information to the dataset or project:                        
if (FALSE) create_coverage(geographic_description = "Description",
                         west_bounding_coordinate = "-160.594000", 
                         east_bounding_coordinate = "-134.104800",
                         north_bounding_coordinate = "71.238300", 
                         south_bounding_coordinate = "67.865000",
                         begin_date = "1980-01-01", 
                         end_date = "2010-12-31", 
                         taxonomic_coverage = taxonomic_coverage)