Add Methods

Adds the methods metadata elements according to EML standards.

add_method(parent_element, methods_file)

Arguments

parent_element

A list representing the EML project or dataset.

methods_file

A file containing methods information (word or markdown): see create_method

Value

The dataset list or project with methods information appended.

Examples

methods_file = word_example("methods-template.docx")

dataset <- list() %>%
    add_method(methods_file)
dataset
#> $methods
#> $methods$sampling
#> NULL
#> 
#> $methods$methodStep
#> $methods$methodStep$instrumentation
#> character(0)
#> 
#> $methods$methodStep$software
#> NULL
#> 
#> $methods$methodStep$description
#> $methods$methodStep$description$section
#> $methods$methodStep$description$section[[1]]
#> [1] "<title>General Protocols</title>\n<para>\n      Field methods. All experiments will be carried out in the\n      greenhouse at Harvard Forest. We have developed an instrumentation\n      system that allows us to collect continuous dissolved [O2]\n      measurements: dedicated micro-probes (DO-166MT; Lazar Research\n      Laboratories: http://www.shelfscientific.com/) connected to\n      multiplexers and data loggers (AM16/32B multiplexer, CR-1000\n      datalogger and control system [Campbell Scientific:\n      http://www.cambellsci.com]). The initial ecosystem composition in\n      all experimental plants will be standardized by seeding each\n      pitcher with a 10-ml inoculum of liquid collected from pitchers\n      growing at Tom Swamp. In all experiments, prey will be supplied to\n      pitchers as standardized aliquots of dried and finely ground\n      bald-faced hornets (Dolichovespula maculata; Hymenoptera:\n      Vespidae), which we collect in quantity throughout New England.\n      Both hornets and ants (the latter are the dominant prey of S.\n      purpurea) are hymenoptera, and have nearly identical C:N ratios\n      (hornets: 3.97; common bog-dwelling ants [Tapinoma sessile and\n      Myrmica lobifrons]: 3.37), but on average hornets have greater\n      than 100 times the dry mass of these ants, and are easier to\n      collect and process as a standardized food source. Additions of\n      prey, either as large \"pulses\" or chronic\n      \"presses\" are analogous to the enrichment and\n      eutrophication that occur in aquatic \"green\" food webs\n      in which phytoplankton abundance is boosted through addition of\n      limiting nutrients. In \"brown\" food webs such as the\n      Sarracenia microecosystem, detritus - not primary production - is\n      at the base of the web, and our treatments boost this material as\n      would happen through increases in arthropod prey capture78 or\n      through nitrogen-enriched precipitation.\n    </para>\n<para>\n      Proteomic analysis. Proteomic profiles of microbial communities\n      are determined after separating the microbial fraction from the\n      pitcher fluid, prey, and other detritus. The microbial\n      \"pellet\" is subjected to SDS-PAGE (sodium dodecyl\n      sulfate polyacrylamide gel) electrophoresis; bands are cut out and\n      digested in-gel with trypsin. Tryptic peptides are subjected to\n      LC-MS/MS (liquid chromatography tandem mass spectrometry) for\n      peptide and protein identification. Absolute abundance of peptides\n      and proteins are quantified using AQUA (Absolute QUAntification)\n      analysis109.\n    </para>"
#> 
#> $methods$methodStep$description$section[[2]]
#> [1] "<title>Specific Experiments</title>\n<para>\n      Experiment #1. Effects of nutrient enrichment on state changes and\n      [O2] profiles. This experiment alters nutrient enrichment rates to\n      characterize the [O2] profile and the transition to the anaerobic\n      state. The experimental design is a one-way layout with 5\n      treatment groups: one control (no enrichment) and 4 enrichment\n      levels (0.125, 0.25, 0.5, 1.0 mg prey added ml-1 d-1). One plant\n      is assigned to each treatment group, and the entire set is\n      replicated 6 times over successive weeks. [O2] is monitored\n      continuously for 4 days to characterize state changes and tipping\n      points under different enrichment rates. This experiment tracks a\n      continuous [O2] profile but does not include proteomic analysis.\n      The purpose of Experiment #1 is to identify an enrichment rate E\n      that generates a long pre-tipping period before transition time T\n      to the anaerobic state. This enrichment rate will be used in\n      Experiments #2 - #4.\n    </para>\n<para>\n      Experiment #2. Identification of early intervention time and\n      characterization of aerobic and anaerobic proteomes. This\n      experiment will use the single enrichment rate E determined from\n      Experiment #1 and impose different intervention times I at which\n      nutrient enrichment will be terminated. Thus, this experiment will\n      identify the latest time I* at which it is possible to intervene\n      and stop the transition to the anaerobic state by halting\n      enrichment. The [O2] profile will again be monitored continuously\n      over 10 days to measure the state of the system. From Experiment\n      #1, the transition time T to the anaerobic state with no\n      intervention will be known. We will use one control group (no prey\n      addition) and ten levels of intervention time (all with the same\n      enrichment rate E) as a proportion of T (0.1, 0.2, 0.3, 0.4, 0.5,\n      0.6, 0.7, 0.8, 0.9, 1.0). Six plants will be assigned randomly to\n      each of the 11 treatments in a randomized one-way layout and [O2]\n      profiles will be monitored continuously. In addition to the [O2]\n      profiles, we will also characterize the protein profiles of\n      aerobic and anaerobic pitchers in all 11 treatment groups at the\n      end of the experiment.\n    </para>\n<para>\n      After the plants are harvested, we will create proteomic profiles\n      of the predominantly bacterial portion (centrifuged pellet) of the\n      pitcher fluid from each plant, as described in General Protocols.\n      Thus, 66 separate pellet-fraction samples will be analyzed by\n      SDS-PAGE. After examining the SDS-PAGE profiles, approximately ten\n      proteins that show dynamic patterns consistent with state change\n      and five that do not change will be cut from the gel, subjected to\n      in-gel tryptic digestion and a portion of the tryptic peptides\n      will be analyzed by LC-MS/MS. Using these data, we will choose\n      three identified peptides from each protein for peptide synthesis\n      such that each synthesized peptide contains stable isotope labels\n      (AQUA peptides) to distinguish them by mass from the native\n      peptides. We will then quantify all 45 of the native peptides from\n      the original samples using a known amount of each AQUA peptide\n      spiked into the tryptic digest. The AQUA analysis of proteins that\n      do not show changes will be used for normalization between\n      samples. These data will be used to independently identify the\n      current state of the system and forecast the time-to-transition.\n    </para>\n<para>\n      We will use Sequest searches for initial identification of\n      peptides; relevant scores including Xcorr and Delta Cn values will\n      be given for each peptide. Other peptides will be identified by de\n      novo sequencing using PepNovo; all PepNovo scores will likewise be\n      given including any N- or C-terminal gaps. Mass error in ppm will\n      be reported for each precursor ion. We will use standard\n      multivariate analysis to search for distinctive proteomic\n      profiles114 that characterize aerobic and anaerobic ecosystems, as\n      well as ecosystems that developed with and without inputs of\n      photosynthetic O2 and plant metabolites.\n    </para>\n<para>\n      Experiment #3. Identification of diagnostic proteins. Using\n      Experiments #1 and 2, we will have identified an enrichment rate E\n      with a long pre-tipping period and an intervention time I* before\n      which mitigation and termination of enrichment will prevent\n      eutrophication. Experiment #3 will characterize the mean and\n      variance of the protein profile before and after I*. We are\n      especially interested in identifying proteins that increase\n      rapidly in abundance (or variance) well before the onset of\n      flickering in [O2] and before the transition time T from the\n      aerobic to the anaerobic state.\n    </para>\n<para>\n      A cohort of 100 plants all will be fed at rate E (determined from\n      Experiment #1), with intervention time I* determined from\n      Experiment #2, although no intervention will be used in this\n      \"press\" experiment so that we can contrast proteins\n      before and after the state change. At seven times before I* and\n      three times after I<emphasis>, we will harvest 10 randomly chosen\n      plants. At each prescribed harvest time, we will measure [O2] and\n      collect samples from each plant for proteomic screening using both\n      SDS-PAGE and AQUA analysis. This experiment will identify proteins\n      that rise quickly in abundance during the pre- I</emphasis> period\n      and can be used as early indicators of a future tipping point.\n      Because different plants will be harvested at each time period,\n      this is a one-way ANOVA design, with pre-and post- I* a priori\n      contrasts. A randomization test will be used to determine whether\n      variances in protein expression differ through time. During these\n      analyses we will use the data from the AQUA peptides and from\n      known amounts of protein standards, such as bovine serum albumin,\n      to approximate the amount of protein in a given coomassie-stained\n      SDS-PAGE gel band. The reason for doing this is to provide a fast\n      \"real-time\" assay based just on expression in the\n      SDS-PAGE. This rapid assay will be used in Experiment #4.\n    </para>\n<para>\n      Experiment # 4. Proof-of-application. This experiment will provide\n      a benchmark test of our methods and their ability to correctly\n      identify tipping points. A cohort of 100 plants will each be\n      fitted with [O2] probes and started on the enrichment regime. Two\n      times per day, we will collect 3 plants each, pool their contents,\n      and conduct a rapid screen in the lab with SDS-PAGE for the\n      diagnostic proteins that were identified in Experiment #3. We will\n      use the protein expression in the gel to delineate an\n      \"early\" and a \"late\" mitigation strategy. As\n      soon as diagnostic proteins measured in the SDS-gels are at\n      abundances that signal we are at 0.5×I* - approximately one-half\n      of the way to the latest intervention time - we will randomly\n      select one third of the remaining plants for mitigation and\n      termination of enrichment (the \"early\" mitigation\n      strategy). We will continue to harvest plants from the remainder\n      of the cohort and monitor proteins. As soon as diagnostic proteins\n      signal we are at 0.75 times I*, we will randomly select one half\n      of the remaining plants for mitigation and termination of\n      enrichment (the \"late\" mitigation strategy). The\n      remaining plants (approximately one sixth to one third of the\n      original cohort) will continue to be enriched. We will monitor\n      [O2] in all 3 groups (no-mitigation control, early mitigation,\n      late mitigation) until all plants reach a new [O2] equilibrium. If\n      the protein markers are successful, the proportion of food webs\n      that remain aerobic will be significantly higher in the two\n      mitigation treatments than in the no-mitigation control.\n    </para>"
#> 
#> 
#> $methods$methodStep$description$para
#> list()
#> 
#> 
#> 
#>