Create Data Table

Creates the data table elements according to EML standards.

create_datatable(
  filepath,
  attribute_info,
  datatable_description,
  datatable_url = NULL,
  dataset_methods = NULL,
  additional_info = NULL
)

Arguments

filepath

Name of the file you would like added to the datatable element

attribute_info

File path to the attribute sheet that contains metadata describing all attributes for your datatable.

datatable_description

Short description of contents of datatable you are adding

datatable_url

(Optional) A URL that links to the datatable you are adding (ex: dataset on AWS public bucket or https://raw.githubusercontent.com/)

dataset_methods

(Optional) metadata describing methods if there are methods that are specific to a datatable

additional_info

(Optional) additional metadata info if there is additional metadata information that is important to the datatable

Value

A data table element listed in a format to be added to a dataset element and turned into EML

Details

Create Data Table

Examples

filepath =  system.file("extdata", "Banet-Example", "data", "enclosure-study-growth-rate-data.csv", package = "EMLaide", mustWork = TRUE)
attribute_info = system.file("extdata", "Banet-Example", "metadata", "enclosure-study-growth-rates-metadata.xlsx", package = "EMLaide", mustWork = TRUE)
datatable_description = "Growth Rates - Enclosure Study"

if (FALSE) create_datatable(filepath, attribute_info, datatable_description)