Create Raster Element

This function creates all the required elements for a spatialRaster section of an ELM document.

create_raster(
  file_name,
  file_description,
  attribute_list,
  physical,
  spatial_reference,
  horizontal_accuracy,
  vertical_accuracy,
  cell_size_x,
  cell_size_y,
  number_of_bands,
  raster_origin,
  rows,
  columns,
  verticals,
  cell_geometry
)

Arguments

file_name
The name of your raster file

file_description
A short description of your file

attribute_list
A list of all the columns on your datatable. Use create_attribute to generate this attribute list and see the documentation for create_attribute for more information.

physical
A list of the physical descriptions of your file. Use create_physical to generate this physical list and see the documentation for create_physical for more information.

spatial_reference
This is the name of a predefined coordinate system. The acceptable values for spatial_reference can be found at EDI Schema.

horizontal_accuracy
The accuracy of horizontal locational measurements within the data.

vertical_accuracy
The accuracy of vertical locational measurements within the data.

cell_size_x
The width of the cell in the x direction.

cell_size_y
The width of the cell in the y direction.

number_of_bands
The number of bands in the image.

raster_origin
The corner location of the pixel it should have the minimum x and y values

rows
Max number of raster object along the y-axis

columns
Max number of raster objects along the x-axis

verticals
Max number of raster objects along the z-axis

cell_geometry
Geometric representation of the cell's content

parent_element
A list to append spatialRaster to

Value

A list that contains all the required elements of the spatialRaster section of an EML document.

Details

Create Raster Element

Examples

if (FALSE) {
create_raster(file_name = "Rasterfiles.tif" ,
              file_description = "A Raster File",
              attribute_list =  create_attribute(attribute_name = "Yrs", attribute_label = "Years", 
                                              attribute_definition = "Calendar year of the observation from years 1990 - 2010.", 
                                              storage_type = EMLaide::storage_type$date, 
                                              measurement_scale = EMLaide::measurement_scale$dateTime, 
                                              date_time_format = "YYYY",
                                              date_time_precision = "1", minimum = "1993", maximum = "2003"),
              physical = create_physical("Rasterfiles.tif"),
              spatial_reference = "NAD_1983_StatePlane_California_I_FIPS_0401",
              horizontal_accuracy = "No Information",
              vertical_accuracy =  "No Information",
              cell_size_x = "25",
              cell_size_y = "25",
              number_of_bands = "10",
              raster_origin = "Upper Left",
              rows = "200",
              columns = "6",
              verticals = "1",
              cell_geometry = "pixel")}